feat: arbitration card in global settings GUI with subscription gating
- Relocate global settings API (handlerMuxSet/SwitchGlobalSettings,
  setRepmanSetting, switchRepmanSetting, setServerSetting,
  switchServerSetting) from api_cluster.go to new api_global_settings.go
- New Arbitration accordion card in ClustersGlobalSettings: enable
  switch, arbitrator address, secret, unique id, peer hosts, failed
  master script, read timeout, and Reset Arbitration (moved from
  Global Settings card)
- Enabling arbitration-external requires IsEligibleForArbitration()
  (registered Cloud18 user with support/support-services/partner plan),
  enforced in API and mirrored in the GUI
- arbitration-external-secret set-case now base64-decodes and updates
  Conf.Secrets (mail-smtp-password pattern) so it is saved encrypted
  in datadir and reloaded on restart
- Fix trailing-space case labels that made settings unreachable via
  API: arbitration-external, api-port, graphite-blacklist,
  graphite-metrics, graphite-carbon-pprof-port, http-bind-address,
  http-use-react, monitoring-save-config
- Change default arbitration-external-hosts to
  https://arbitrator.cloud18.io (was 88.191.151.84:80)
- Document runtime apply case and secret procedure in
  ADDING_GLOBAL_SETTINGS.md
